Fix webpack/`use client` issue
