+++ frontend/app/page.tsx
@@ -1,17 +1,13 @@
-"use client";
-
-import { useState } from "react";
-import Login from "./components/Login";
+import Link from "next/link";
 /**
  * Home component, returns main content
  * @returns {TSX}
  */
 export default function Home() {
-  const [count, setCount] = useState(0);
   return (
     <div>
       <h1>RMoods</h1>
-      <Login />
+      <Link href={"/login"}>Go to Login</Link>
     </div>
   );
 }
